No need to read all the partition data and add to buffer, we will read only one chunk and then output for rows between unbounded preceding and current row.

for sql

```
select count(lo_orderkey), count(lo_partkey), count(lo_custkey), count(lo_suppkey), count(lo_quantity), count(lo_discount), count(lo_shipmode), count(_num) from 
(select lo_orderkey, lo_partkey, lo_custkey, lo_suppkey, lo_quantity, lo_discount, lo_shipmode, row_number() over(order by lo_partkey) as _num from lineorder) t1;
```

parallel = 4

before opt: TheLastFragment PeakMem(4.55G), time(25.257s)

```
    Fragment 0:
      Instance a570aad6-d009-11ec-8733-66660a909346 (host=TNetworkAddress(hostname:172.26.92.195, port:10061)):(Active: 25s247ms[25247329630ns], % non-child: 0.00%)
         - AverageThreadTokens: 4607182418800017400.00
         - MemoryLimit: 18.63 GB
         - PeakMemoryUsage: 4.55 GB
         - RowsProduced: 1
 ```

after opt: TheLastFragment PeakMem(782.66M), time(24.614s)

```
      Instance ab4b09fe-d008-11ec-8733-66660a909346 (host=TNetworkAddress(hostname:172.26.92.195, port:10061)):(Active: 24s602ms[24602663842ns], % non-child: 0.00%)
         - AverageThreadTokens: 4607182418800017400.00
         - MemoryLimit: 18.63 GB
         - PeakMemoryUsage: 782.66 MB
         - RowsProduced: 1
```