@trueeyu 2022-05-12
run starrocks_fe_unittest